API_Server/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
firebase-admin==6.2.0
python-dotenv==1.0.0
google-cloud-aiplatform==1.71.1
google-ai-generativelanguage==0.4.0
google-generativeai==0.3.2
vertexai==1.71.1
Pillow==10.0.1
requests==2.31.0
gunicorn==21.2.0
PyJWT==2.9.0
flask-restx==1.3.0
PyMySQL==1.1.0
psycopg2-binary==2.9.9
cryptography==42.0.0
google-cloud-storage==2.10.0
google-auth==2.23.4
google-cloud-core==2.3.3